October update
Started messing around with stat function for more file information, and
coded a function to parse the data it gives.

IDE is haunted currently, no matter what is actually included in the
code, it still runs the stat function.

Removed stat for right now and added the replacement function Stat_e.

Backup and restore arguments 'b' and 'b r' respectively